+++ src/core/primitives/js_interop_primitives.js
@@ -11,14 +11,8 @@ export const jsInteropPrimitives = {
     // We expect the Scheme side to define the record type, then pass the constructor here
     'register-js-object-record': (ctor) => {
         registerJsObjectRecord(ctor);
-    },
-
-    // Generic Accessors for the js-object (which are just JS objects)
-    'js-ref': (obj, key) => {
-        return obj[key];
-    },
-
-    'js-set!': (obj, key, val) => {
-        obj[key] = val;
     }
+    // Note: js-ref and js-set! are defined in extras/primitives/interop.js
+    // with proper jsToScheme conversion. Do not duplicate them here!
 };
+